fix(автонастройка): прокси игнорировался, а Telegram id не влезал в int32
Две поломки, найденные проверкой аккаунта НА ПРОДЕ живым кодом услуги. Обе
молчаливые: ни ошибки, ни строки в логе — просто тихо неверное поведение.

1. ПРОКСИ НЕ ПРИМЕНЯЛСЯ ВОВСЕ. Telethon 1.43 берёт прокси через python-socks,
   а не PySocks, и `proxy_type` там СТРОКА ('socks5'), а не константа PySocks.
   Пакета на проде не было, и Telethon писал в лог «proxy argument will be
   ignored» и шёл НАПРЯМУЮ: узбекский номер подключался с российского IP
   194.156.119.17 — ровно то, за что Telegram блокирует аккаунты, и мы бы
   узнали об этом только по факту бана.

   Промежуточная попытка отдать голый URL строкой тоже не работает — Telethon
   отвечает «Proxy of unknown format: <class 'str'>», ему нужен словарь.

   Проверено после починки: сервер выходит с 194.156.119.17, аккаунт через
   прокси — с 170.168.109.18. @SpamBot отвечает «no limits are applied».

2. NULLIF($5,0) РОНЯЛ ЗАПИСЬ СОСТОЯНИЯ. asyncpg выводит тип литерала 0 как
   int32, а Telegram id давно длиннее (8741578822 > 2^31): «invalid input for
   query argument $5: value out of int32 range». То есть проверка аккаунта
   отрабатывала, но результат в базу не ложился. Лечится приведением ::bigint;
   поправлено в обоих местах — задача health_check и админский эндпоинт.

+++ backend/app/api/admin_tg_setup.py
@@ -192,7 +192,10 @@ async def check_account(account_id: int,
         """UPDATE tg_setup_accounts
               SET health=$2, health_note=$3, health_checked_at=NOW(),
                   username=COALESCE(NULLIF($4,''), username),
-                  tg_user_id=COALESCE(NULLIF($5,0), tg_user_id), updated_at=NOW()
+                  -- ⚠️ ::bigint ОБЯЗАТЕЛЕН — см. пояснение в tasks/tg_setup.py:
+                  -- без него asyncpg считает 0 за int32, а Telegram id длиннее.
+                  tg_user_id=COALESCE(NULLIF($5::bigint, 0::bigint), tg_user_id),
+                  updated_at=NOW()
             WHERE id=$1""",
         account_id, health.state, health.note, health.username, health.tg_user_id,
     )
